requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2021.5.30
chardet==3.0.4
Click==7.0
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
djangorestframework==3.13.1
filelock==3.0.12
frappe-bench==5.4.1
gitdb2==2.0.6
GitPython==2.1.15
honcho==1.0.1
idna==2.8
Jinja2==2.11.3
joblib==1.1.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.20.3
pandas==1.1.5
python-crontab==2.4.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.22.0
scikit-learn==1.0.1
scipy==1.8.0
semantic-version==2.8.2
six==1.15.0
smmap==4.0.0
smmap2==3.0.1
sqlparse==0.4.1
threadpoolctl==3.1.0
tzdata==2021.5
urllib3==1.25.11
virtualenv==20.4.3
